feat: 优化代码结构
