+++ packages/next-wxt/entrypoints/sidepanel/utils/snapshotOperations.ts
@@ -1,91 +1,8 @@
-// 快照操作工具
-// 参考 chrome-devtools-mcp 的操作方式
-// 提供点击、输入等公共操作方法
-
-import type { ElementHandle } from 'puppeteer-core'
 import { SnapshotManager } from './snapshotManager'
-
-/**
- * 延迟函数
- */
-function delay(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms))
-}
-
-/**
- * 等待 DOM 稳定
- * 参考 chrome-devtools-mcp 的 waitForStableDom
- */
-async function waitForStableDom(page: any, timeout = 5000): Promise<void> {
-  try {
-    // 简单的延迟等待，100ms 后认为 DOM 稳定
-    // 更精确的实现需要使用 CDP 的 DOM 事件监听
-    await delay(100)
-
-    // 可选：等待网络空闲
-    try {
-      await page.waitForLoadState?.('networkidle', { timeout: Math.min(timeout, 2000) }).catch(() => {
-        // 如果超时，忽略错误
-      })
-    } catch {
-      // 如果 waitForLoadState 不存在，忽略
-    }
-  } catch (error: any) {
-    // 如果等待失败，忽略错误，继续执行
-    console.warn('等待 DOM 稳定失败:', error)
-  }
-}
-
-/**
- * 等待导航完成
- * 参考 chrome-devtools-mcp 的 waitForNavigation
- */
-async function waitForNavigation(page: any, timeout = 30000): Promise<void> {
-  try {
-    // 尝试等待导航完成
-    await Promise.race([
-      page.waitForNavigation?.({ timeout, waitUntil: 'networkidle0' }).catch(() => {
-        // 如果没有导航，返回
-        return Promise.resolve()
-      }),
-      delay(Math.min(timeout, 5000)).then(() => {
-        // 5秒后如果没有导航，返回
-        return Promise.resolve()
-      })
-    ])
-  } catch (error: any) {
-    // 如果导航超时或没有导航，忽略错误
-    console.warn('等待导航失败:', error)
-  }
-}
-
-/**
- * 执行操作并等待事件
- * 参考 chrome-devtools-mcp 的 waitForEventsAfterAction
- */
-async function waitForEventsAfterAction(page: any, action: () => Promise<void>): Promise<void> {
-  try {
-    // 启动导航监听（如果有）
-    const navigationPromise = waitForNavigation(page).catch(() => {
-      // 如果没有导航，忽略错误
-    })
-
-    // 执行操作
-    await action()
-
-    // 等待导航完成（如果有）
-    await navigationPromise
-
-    // 等待 DOM 稳定
-    await waitForStableDom(page)
-  } catch (error: any) {
-    throw new Error(`操作后等待事件失败: ${error.message}`)
-  }
-}
+import { delay, waitForEventsAfterAction } from './utils'
 
 /**
  * 通过 UID 点击节点
- * 参考 chrome-devtools-mcp 的 click 工具
  * @param manager 快照管理器
  * @param uid 节点 UID
  * @param options 点击选项
